# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Anby Wiki API
  version: "0.1.0"
  description: |
    Anby Wiki HTTP API 权威契约（OpenAPI 3.1）。
    本文件是所有接口行为的唯一权威定义：错误模型、分页、幂等头与请求 ID 在此统一定义。
    前端客户端由 OpenAPI Generator 从本文件生成，禁止手写 URL/DTO。
    Go 侧 handler 手写实现，由 Contract Test 保证与本文件一致。
  contact:
    name: Anby Wiki
...
servers:
  - url: http://localhost:3000
    description: 本地 Web 入口（Next.js 将 /api/* 转发到 Go API）
  - url: http://localhost:8080
    description: 本地直连 Go API
...
paths:
  /api/v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: 注册本地账号并建立会话
      description: >-
        用户名和邮箱均不区分大小写且全站唯一。首个本地账号获得管理员角色，
        后续账号默认获得编辑者角色；注册默认关闭，部署方须通过
        AUTH_REGISTRATION_ENABLED 显式开启。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: 注册成功，设置 HttpOnly session cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: 使用用户名或邮箱和密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: 登录成功，设置 HttpOnly session cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: 获取当前登录 Actor
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 当前登录会话
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: 吊销当前服务端 session 并清除 cookie
      security:
        - sessionCookie: []
      responses:
        "204":
          description: 已退出；无活动 session 时同样成功
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/import-jobs:
    get:
      operationId: listImportJobs
      tags: [imports]
      summary: 分页列出当前账号的导入任务
      description: 按创建时间倒序返回，刷新或重新登录后仍可恢复完整导入队列。
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [queued, running, succeeded, failed, cancelled]
      responses:
        "200":
          description: 当前账号的导入任务
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJobListPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      operationId: createImportJob
      tags: [imports]
      summary: 幂等创建来源导入任务
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateImportJobRequest" }
      responses:
        "201":
          description: 已排队的导入任务
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: 列出当前站点的媒体资产
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - name: kind
          in: query
          schema:
            type: string
            enum: [image, video, other]
      responses:
        "200":
          description: 资产目录游标分页结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetListPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/Internal" }
    post:
      operationId: uploadAsset
      tags: [assets]
      summary: 上传媒体资产并创建不可变版本
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
      responses:
        "200":
          description: 内容去重命中，返回既有资产版本
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "201":
          description: 新资产或新版本
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/Internal" }
  /api/v1/assets/revisions/{revision_id}:
    get:
      operationId: getAssetRevision
      tags: [assets]
      summary: 读取不可变资产版本元数据
      parameters:
        - $ref: "#/components/parameters/AssetRevisionId"
      responses:
        "200":
          description: 资产版本元数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetRevision" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/assets/revisions/{revision_id}/content:
    get:
      operationId: getAssetRevisionContent
      tags: [assets]
      summary: 流式读取不可变资产版本内容
      parameters:
        - $ref: "#/components/parameters/AssetRevisionId"
      responses:
        "200":
          description: 带 immutable 缓存头的原始资产内容
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "304":
          description: ETag 未变化
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/Internal" }
  /api/v1/sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: 分页浏览逻辑来源目录
      description: 只查询来源元数据，不读取 SourceChunk 正文。
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - name: source_type
          in: query
          schema:
            type: string
            enum: [webpage, pdf, book, image, video, api, database]
        - name: q
          in: query
          description: 标题、作者或发布者的元数据搜索，最长 200 字符。
          schema:
            type: string
            maxLength: 200
      responses:
        "200":
          description: 来源目录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceSourceListPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      operationId: createSource
      tags: [sources]
      summary: 登记逻辑来源
      description: URL 会经领域服务规范化并幂等关联 ExternalResource。
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEvidenceSourceRequest" }
      responses:
        "201":
          description: 已创建来源
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceSource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/sources/{id}:
    get:
      operationId: getSource
      tags: [sources]
      summary: 读取来源与规范化外部资源
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 来源详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceSourceDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/sources/{id}/versions:
    get:
      operationId: listSourceVersions
      tags: [sources]
      summary: 分页列出来源的不可变版本
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 来源版本目录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceSourceVersionListPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/source-versions/{id}/chunks:
    get:
      operationId: listSourceChunks
      tags: [sources]
      summary: 分页列出不可变来源版本的可定位分片
      description: SourceChunk 正文可能包含来源全文片段，不写入日志。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 来源分片目录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceSourceChunkListPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/citations:
    post:
      operationId: createCitation
      tags: [sources]
      summary: 创建不可变 Citation
      description: 引文若绑定 SourceChunk，必须是该分片文本的真实子串。
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCitationRequest" }
      responses:
        "201":
          description: 已创建 Citation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CitationRecord" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/datasets:
    get:
      operationId: listDatasets
      tags: [datasets]
      summary: 列出站点 Dataset
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Dataset 游标分页结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetListPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createDataset
      tags: [datasets]
      summary: 创建带 JSON Schema 的 Dataset
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDatasetRequest" }
      responses:
        "201":
          description: Dataset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dataset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/datasets/{id}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: 读取 Dataset 定义
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      responses:
        "200":
          description: Dataset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dataset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/datasets/{id}/records:
    get:
      operationId: listDatasetRecords
      tags: [datasets]
      summary: 按稳定默认顺序列出 DatasetRecord
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: DatasetRecord 分页结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetRecordPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createDatasetRecord
      tags: [datasets]
      summary: 创建经 Dataset Schema 校验的记录
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WriteDatasetRecordRequest" }
      responses:
        "201":
          description: DatasetRecord
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/datasets/{id}/views:
    get:
      operationId: listDatasetViews
      tags: [datasets]
      summary: 列出 Dataset 的保存视图
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      responses:
        "200":
          description: DatasetView 列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetViewList" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createDatasetView
      tags: [datasets]
      summary: 创建筛选、排序与分组视图
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDatasetViewRequest" }
      responses:
        "201":
          description: DatasetView
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/dataset-records/{id}:
    put:
      operationId: updateDatasetRecord
      tags: [datasets]
      summary: 更新记录并重新执行 Dataset Schema 校验
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WriteDatasetRecordRequest" }
      responses:
        "200":
          description: DatasetRecord
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/dataset-views/{id}:
    get:
      operationId: getDatasetView
      tags: [datasets]
      summary: 读取保存的 DatasetView
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      responses:
        "200":
          description: DatasetView
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetView" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/dataset-views/{id}/records:
    get:
      operationId: queryDatasetView
      tags: [datasets]
      summary: 执行保存视图的筛选、排序与分组
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: DatasetView 查询结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatasetRecordPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/components:
    get:
      operationId: listWikiComponents
      tags: [components]
      summary: 列出组件注册项
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 组件游标分页结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WikiComponentListPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createWikiComponent
      tags: [components]
      summary: 创建稳定组件注册项
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWikiComponentRequest" }
      responses:
        "201":
          description: 组件
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WikiComponent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/components/{id}:
    get:
      operationId: getWikiComponent
      tags: [components]
      summary: 读取组件注册项
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      responses:
        "200":
          description: 组件
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WikiComponent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/import-jobs/{id}:
    get:
      operationId: getImportJob
      tags: [imports]
      summary: 读取导入任务和各次运行阶段进度
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/GovernanceId"
      responses:
        "200":
          description: 导入任务详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJobDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: anby_session
      description: HttpOnly 同源服务端 session cookie；名称可由部署配置覆盖
...
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[A-Za-z0-9][A-Za-z0-9_.-]{2,31}$"
          description: 不区分大小写；服务端规范化为小写。
        email:
          type: string
          format: email
          maxLength: 254
          description: 不区分大小写。
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
        display_name:
          type: string
          maxLength: 128
          description: 省略时使用规范化用户名。
...
    AuthResult:
      type: object
      required: [actor_id, display_name]
      properties:
        actor_id:
          type: string
          format: uuid
        display_name:
          type: string
...
  responses:
    BadRequest:
      description: 请求格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...